Extract artist method for getall and create get_all_artists method for ArtistController
